Se reemplaza el "store" de la etiqueta por etiqueta y set para cuando se realice un busqueda en VC no se confundan etiquetas iguales de sets distintos

+++ colaDirecciones.h
@@ -3,7 +3,7 @@
 
 #include <stdbool.h>
 
-typedef unsigned long etiquetaDireccion;
+typedef unsigned long long etiquetaDireccion;
 typedef struct celdaC{
     etiquetaDireccion elem;
     struct celdaC *ant;
